functions/load_data_into_rawzone/requirements.txt
wget==3.2
pandas_gbq==0.14.1
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
SQLAlchemy==1.3.23
google-cloud-storage==1.37.1
google-cloud-bigquery-storage==2.4.0
google-cloud-bigquery==2.13.1